requirements.txt
Flask==2.3.3
Pillow==10.1.0
rembg==2.0.57
flask-cors==4.0.0
werkzeug==3.0.1
gunicorn==21.2.0